+++ packages/palette/src/components/StaticCountdownTimer/StaticCountdownTimer.tsx
@@ -3,13 +3,8 @@
 import moment from "moment-timezone"
 import React from "react"
 import { TimeRemaining } from "../"
-import {
-  Flex,
-  ProgressBarTimer,
-  Sans,
-  StackableBorderBox,
-} from "../../elements"
-import { Spacer } from "../../tokens"
+import { Flex, ProgressBarTimer, StackableBorderBox } from "../../elements"
+import { Sans, Spacer } from "../../tokens"
 
 import { TimerIcon } from "../../svgs"
