+++ packages/palette/src/components/TimeRemaining/TimeRemaining.tsx
@@ -1,8 +1,8 @@
 import moment from "moment"
 import React from "react"
 
-import { Sans } from "../../elements"
 import { color } from "../../helpers"
+import { Sans } from "../../tokens"
 
 const pad = (n: number) => n.toString().padStart(2, "0")
