+++ packages/palette/src/elements/index.tsx
@@ -28,4 +28,3 @@ export * from "./Tabs"
 export * from "./TextArea"
 export * from "./Toggle"
 export * from "./Tooltip"
-export * from "./Typography"
