+++ packages/palette/src/tokens/Colors/Colors.tsx
@@ -2,8 +2,8 @@ import React from "react"
 import styled from "styled-components"
 import { BackgroundProps, color, ColorProps } from "styled-system"
 import { Flex } from "../../elements/Flex"
-import { Sans } from "../../elements/Typography"
 import { themeProps } from "../../Theme"
+import { Sans } from "../Typography"
 
 export interface ColorBlockProps extends ColorProps, BackgroundProps {}
