+++ packages/palette/src/tokens/index.tsx
@@ -1,3 +1,4 @@
 export * from "./Colors"
 export * from "./Grid"
 export * from "./Spacer"
+export * from "./Typography"
